requirements.txt
# make sure there were installed by the base
numpy
tensorflow-gpu~=2.3.0

tensorflow-probability~=0.11.1

sklearn
pandas
matplotlib
h5py
Pillow

opencv-python

tqdm